// src/models/types/dispatch-rules/DispatchRuleTypes.ts
export enum DispatchRuleType {
    COMBINED = "COMBINED",
    GROUP = "GROUP",
    VAT_GROUP = "VAT_GROUP",
    TYPE = "TYPE",
    STORE = "STORE",
    BANK_MOVEMENT_TYPE = "BANK_MOVEMENT_TYPE",
    BANK_ACCOUNT = "BANK_ACCOUNT",
    BANK_COUNTERPART = "BANK_COUNTERPART",
    BANK_CAUSAL = "BANK_CAUSAL",
}